@@ -1,14 +1,23 @@
 [tool.ruff]
-target-version = "py311"
+target-version = "py312"
 line-length = 100
-select = ["E", "F", "I", "N", "W", "UP", "ANN", "S", "B", "A", "C4", "T20"]
-ignore = ["ANN101", "ANN102", "S101"]
 
-[tool.ruff.per-file-ignores]
-"tests/*" = ["ANN", "S"]
+[tool.ruff.lint]
+select = ["E", "F", "I", "N", "UP", "S"]
+ignore = [
+    "S101",   # assert used in tests
+    "S104",   # binding to 0.0.0.0 is intentional for containers
+    "N802",   # uppercase property names are intentional (JWT_SECRET etc.)
+    "S105",   # "bearer" is not a hardcoded password
+]
+
+[tool.ruff.lint.per-file-ignores]
+"tests/*" = ["S"]
+"alembic/*" = ["I"]
+"migrations/*" = ["I"]
 
 [tool.mypy]
-python_version = "3.11"
+python_version = "3.12"
 strict = true
 ignore_missing_imports = true
 plugins = ["pydantic.mypy"]
