@@ -5,7 +5,7 @@ Provides session cache, rate limiting, and WebSocket state.
 from __future__ import annotations
 
 import logging
-from typing import AsyncGenerator
+from collections.abc import AsyncGenerator
 
 import redis.asyncio as aioredis
 
