@@ -8,7 +8,7 @@ from __future__ import annotations
 import json
 import time
 import uuid
-from typing import Callable
+from collections.abc import Callable
 
 from fastapi import Request
 from starlette.middleware.base import BaseHTTPMiddleware
