@@ -6,7 +6,7 @@ and per workspace for agent API calls.
 from __future__ import annotations
 
 import time
-from typing import Callable
+from collections.abc import Callable
 
 import redis.asyncio as aioredis
 from fastapi import HTTPException, Request, status
