@@ -1,5 +1,5 @@
+from .auth_service import AuthService
 from .pipeline_service import PipelineService
 from .workspace_service import WorkspaceService
-from .auth_service import AuthService
 
 __all__ = ["PipelineService", "WorkspaceService", "AuthService"]
