@@ -1,3 +1,3 @@
-from .pipeline_worker import start_pipeline_worker, run_pipeline_direct
+from .pipeline_worker import run_pipeline_direct, start_pipeline_worker
 
 __all__ = ["start_pipeline_worker", "run_pipeline_direct"]
